+++ unidade-III/pilhas/src/pilha.h
@@ -8,9 +8,9 @@ typedef  struct pilha Pilha;
 
 Pilha *pilha_cria(void);
 
-void pilha_push(Pilha *p, float v);
+void pilha_push(Pilha *p, int v);
 
-float pilha_pop(Pilha *p);
+int pilha_pop(Pilha *p);
 
 int pilha_vazia(Pilha* p);
